+++ tests/integration/plugins/test_maven_use/from_binaries/binaries_source/java-dep-add/src/main/java/add/Adder.java
@@ -1,0 +1,8 @@
+package org.starcraft.add;
+
+public class Adder {
+  public int add(int value1, int value2){
+    return value1 + value2;
+  }
+}
+